backend/requirements.txt
blinker==1.7.0
boto3==1.34.44
botocore==1.34.44
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.17.0
